+++ tests/Util/Dca/DcaUtilTest.php
@@ -10,7 +10,7 @@ namespace HeimrichHannot\UtilsBundle\Tests\Util\Dca;
 
 use Contao\Controller;
 use HeimrichHannot\UtilsBundle\Tests\AbstractUtilsTestCase;
-use HeimrichHannot\UtilsBundle\Util\DcaUtil\DcaUtil;
+use HeimrichHannot\UtilsBundle\Util\DcaUtil;
 use HeimrichHannot\UtilsBundle\Util\DcaUtil\GetDcaFieldsOptions;
 use PHPUnit\Framework\MockObject\MockBuilder;
